Add files via upload
Its a fun game. The system is the opponent. 
it generates a N-digited number of your choice and allows you to guess it. 
if the digit of your number are in the system number but not in the same position as in it ,
then it increases Orange value by 1. If digit of your number in as same place as in It then it increases the Green number by 1
This goes on till you guess the correct number .